Update README.md
I created a data story about Covid-19 situation in the Philippines from year 2020 to 2022. I created individual charts to compare the data about Covid Cases, Covid Deaths and Covid Vaccinated by month and year in order to see the difference and also I include a Pivot tables to filter and give the neccessary data for the specific date and time.